dashboard-api/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pandas==2.1.3
slowapi==0.1.9
isodate==0.6.1
